+++ source/lib/mlops_orchestrator_stack.py
@@ -29,7 +29,6 @@ from aws_solutions_constructs import aws_apigateway_lambda
 from lib.conditional_resource import ConditionalResources
 from lib.blueprints.byom.pipeline_definitions.helpers import (
     suppress_s3_access_policy,
-    apply_secure_bucket_policy,
     suppress_lambda_policies,
     suppress_sns,
 )
@@ -51,12 +50,18 @@ from lib.blueprints.byom.pipeline_definitions.iam_policies import (
 from lib.blueprints.byom.pipeline_definitions.configure_multi_account import (
     configure_multi_account_parameters_permissions,
 )
-from lib.blueprints.byom.pipeline_definitions.sagemaker_model_registry import create_sagemaker_model_registry
-from lib.blueprints.byom.pipeline_definitions.cdk_context_value import get_cdk_context_value
+from lib.blueprints.byom.pipeline_definitions.sagemaker_model_registry import (
+    create_sagemaker_model_registry,
+)
+from lib.blueprints.byom.pipeline_definitions.cdk_context_value import (
+    get_cdk_context_value,
+)
 
 
 class MLOpsStack(core.Stack):
-    def __init__(self, scope: core.Construct, id: str, *, multi_account=False, **kwargs) -> None:
+    def __init__(
+        self, scope: core.Construct, id: str, *, multi_account=False, **kwargs
+    ) -> None:
         super().__init__(scope, id, **kwargs)
 
         # Get stack parameters:
@@ -74,16 +79,24 @@ class MLOpsStack(core.Stack):
         allow_detailed_error_message = pf.create_detailed_error_message_parameter(self)
 
         # Conditions
-        git_address_provided = cf.create_git_address_provided_condition(self, git_address)
+        git_address_provided = cf.create_git_address_provided_condition(
+            self, git_address
+        )
 
         # client provided an existing S3 bucket name, to be used for assets
-        existing_bucket_provided = cf.create_existing_bucket_provided_condition(self, existing_bucket)
+        existing_bucket_provided = cf.create_existing_bucket_provided_condition(
+            self, existing_bucket
+        )
 
         # client provided an existing Amazon ECR name
-        existing_ecr_provided = cf.create_existing_ecr_provided_condition(self, existing_ecr_repo)
+        existing_ecr_provided = cf.create_existing_ecr_provided_condition(
+            self, existing_ecr_repo
+        )
 
         # client wants the solution to create model registry
-        model_registry_condition = cf.create_model_registry_condition(self, create_model_registry)
+        model_registry_condition = cf.create_model_registry_condition(
+            self, create_model_registry
+        )
 
         # S3 bucket needs to be created for assets
         create_new_bucket = cf.create_new_bucket_condition(self, existing_bucket)
@@ -101,13 +114,18 @@ class MLOpsStack(core.Stack):
             encryption=s3.BucketEncryption.S3_MANAGED,
             block_public_access=s3.BlockPublicAccess.BLOCK_ALL,
             versioned=True,
+            enforce_ssl=True,
         )
 
-        # Apply secure transfer bucket policy
-        apply_secure_bucket_policy(access_logs_bucket)
+        # remove ACL to add bucket policy instead
+        access_logs_bucket.node.default_child.add_deletion_override(
+            "Properties.AccessControl"
+        )
 
         # This is a logging bucket.
-        access_logs_bucket.node.default_child.cfn_options.metadata = suppress_s3_access_policy()
+        access_logs_bucket.node.default_child.cfn_options.metadata = (
+            suppress_s3_access_policy()
+        )
 
         # Import user provide S3 bucket, if any. s3.Bucket.from_bucket_arn is used instead of
         # s3.Bucket.from_bucket_name to allow cross account bucket.
@@ -118,7 +136,9 @@ class MLOpsStack(core.Stack):
         )
 
         # Create the resource if existing_bucket_provided condition is True
-        core.Aspects.of(client_existing_bucket).add(ConditionalResources(existing_bucket_provided))
+        core.Aspects.of(client_existing_bucket).add(
+            ConditionalResources(existing_bucket_provided)
+        )
 
         # Import user provided Amazon ECR repository
 
@@ -126,7 +146,9 @@ class MLOpsStack(core.Stack):
             self, "ClientExistingECRReo", existing_ecr_repo.value_as_string
         )
         # Create the resource if existing_ecr_provided condition is True
-        core.Aspects.of(client_erc_repo).add(ConditionalResources(existing_ecr_provided))
+        core.Aspects.of(client_erc_repo).add(
+            ConditionalResources(existing_ecr_provided)
+        )
 
         # Creating assets bucket so that users can upload ML Models to it.
         assets_bucket = s3.Bucket(
@@ -137,10 +159,8 @@ class MLOpsStack(core.Stack):
             server_access_logs_bucket=access_logs_bucket,
             server_access_logs_prefix="assets_bucket_access_logs",
             block_public_access=s3.BlockPublicAccess.BLOCK_ALL,
+            enforce_ssl=True,
         )
-
-        # Apply secure transport bucket policy
-        apply_secure_bucket_policy(assets_bucket)
 
         # Create the resource if create_new_bucket condition is True
         core.Aspects.of(assets_bucket).add(ConditionalResources(create_new_bucket))
@@ -151,6 +171,39 @@ class MLOpsStack(core.Stack):
             client_existing_bucket.bucket_name,
             assets_bucket.bucket_name,
         ).to_string()
+
+        blueprints_bucket_name = "blueprint-repository-" + str(uuid.uuid4())
+        blueprint_repository_bucket = s3.Bucket(
+            self,
+            blueprints_bucket_name,
+            encryption=s3.BucketEncryption.S3_MANAGED,
+            server_access_logs_bucket=access_logs_bucket,
+            server_access_logs_prefix=blueprints_bucket_name,
+            block_public_access=s3.BlockPublicAccess.BLOCK_ALL,
+            versioned=True,
+            enforce_ssl=True,
+        )
+
+        # add S3 Server Access Logs Policy
+        access_logs_bucket.add_to_resource_policy(
+            iam.PolicyStatement(
+                effect=iam.Effect.ALLOW,
+                actions=["s3:PutObject"],
+                principals=[iam.ServicePrincipal("logging.s3.amazonaws.com")],
+                resources=[
+                    f"{access_logs_bucket.bucket_arn}/*",
+                ],
+                conditions={
+                    "ArnLike": {
+                        "aws:SourceArn": [
+                            f"arn:{core.Aws.PARTITION}:s3:::{assets_s3_bucket_name}",
+                            blueprint_repository_bucket.bucket_arn,
+                        ]
+                    },
+                    "StringEquals": {"aws:SourceAccount": core.Aws.ACCOUNT_ID},
+                },
+            )
+        )
 
         # Creating Amazon ECR repository
         ecr_repo = ecr.Repository(self, "ECRRepo", image_scan_on_push=True)
@@ -172,27 +225,18 @@ class MLOpsStack(core.Stack):
             ecr_repo.repository_arn,
         ).to_string()
 
-        blueprints_bucket_name = "blueprint-repository-" + str(uuid.uuid4())
-        blueprint_repository_bucket = s3.Bucket(
-            self,
-            blueprints_bucket_name,
-            encryption=s3.BucketEncryption.S3_MANAGED,
-            server_access_logs_bucket=access_logs_bucket,
-            server_access_logs_prefix=blueprints_bucket_name,
-            block_public_access=s3.BlockPublicAccess.BLOCK_ALL,
-            versioned=True,
-        )
-        # Apply secure transport bucket policy
-        apply_secure_bucket_policy(blueprint_repository_bucket)
-
         # create sns topic and subscription
         mlops_notifications_topic = sns.Topic(
             self,
             "MLOpsNotificationsTopic",
         )
-        mlops_notifications_topic.node.default_child.cfn_options.metadata = suppress_sns()
+        mlops_notifications_topic.node.default_child.cfn_options.metadata = (
+            suppress_sns()
+        )
         mlops_notifications_topic.add_subscription(
-            subscriptions.EmailSubscription(email_address=notification_email.value_as_string)
+            subscriptions.EmailSubscription(
+                email_address=notification_email.value_as_string
+            )
         )
 
         # grant EventBridge permissions to publish messages
@@ -216,9 +260,13 @@ class MLOpsStack(core.Stack):
 
         # creating SageMaker Model registry
         # use the first 8 characters as a unique_id to be appended to the model_package_group_name
-        unique_id = core.Fn.select(0, core.Fn.split("-", create_id_function.get_att_string("UUID")))
+        unique_id = core.Fn.select(
+            0, core.Fn.split("-", create_id_function.get_att_string("UUID"))
+        )
         model_package_group_name = f"mlops-model-registry-{unique_id}"
-        model_registry = create_sagemaker_model_registry(self, "SageMakerModelRegistry", model_package_group_name)
+        model_registry = create_sagemaker_model_registry(
+            self, "SageMakerModelRegistry", model_package_group_name
+        )
 
         # only create based on the condition
         model_registry.cfn_options.condition = model_registry_condition
@@ -227,7 +275,9 @@ class MLOpsStack(core.Stack):
         model_registry.node.add_dependency(create_id_function)
 
         # Custom resource to copy source bucket content to blueprints bucket
-        custom_resource_lambda_fn = create_copy_assets_lambda(self, blueprint_repository_bucket.bucket_name)
+        custom_resource_lambda_fn = create_copy_assets_lambda(
+            self, blueprint_repository_bucket.bucket_name
+        )
 
         # grant permission to upload file to the blueprints bucket
         blueprint_repository_bucket.grant_write(custom_resource_lambda_fn)
@@ -246,12 +296,18 @@ class MLOpsStack(core.Stack):
 
         # Cloudformation policy setup
         orchestrator_policy = create_orchestrator_policy(
-            self, pipeline_stack_name, ecr_repo_name, blueprint_repository_bucket, assets_s3_bucket_name
+            self,
+            pipeline_stack_name,
+            ecr_repo_name,
+            blueprint_repository_bucket,
+            assets_s3_bucket_name,
         )
         orchestrator_policy.attach_to_role(cloudformation_role)
 
         # Lambda function IAM setup
-        lambda_passrole_policy = iam.PolicyStatement(actions=["iam:passrole"], resources=[cloudformation_role.role_arn])
+        lambda_passrole_policy = iam.PolicyStatement(
+            actions=["iam:passrole"], resources=[cloudformation_role.role_arn]
+        )
         # create sagemaker layer
         sm_layer = sagemaker_layer(self, blueprint_repository_bucket)
         # make sure the sagemaker code is uploaded first to the blueprints bucket
@@ -261,7 +317,7 @@ class MLOpsStack(core.Stack):
             self,
             "PipelineOrchestration",
             lambda_function_props={
-                "runtime": lambda_.Runtime.PYTHON_3_8,
+                "runtime": lambda_.Runtime.PYTHON_3_9,
                 "handler": "index.handler",
                 "code": lambda_.Code.from_asset("lambdas/pipeline_orchestration"),
                 "layers": [sm_layer],
@@ -278,15 +334,25 @@ class MLOpsStack(core.Stack):
         )
 
         # add lambda suppressions
-        provisioner_apigw_lambda.lambda_function.node.default_child.cfn_options.metadata = suppress_lambda_policies()
+        provisioner_apigw_lambda.lambda_function.node.default_child.cfn_options.metadata = (
+            suppress_lambda_policies()
+        )
 
-        provision_resource = provisioner_apigw_lambda.api_gateway.root.add_resource("provisionpipeline")
+        provision_resource = provisioner_apigw_lambda.api_gateway.root.add_resource(
+            "provisionpipeline"
+        )
         provision_resource.add_method("POST")
-        status_resource = provisioner_apigw_lambda.api_gateway.root.add_resource("pipelinestatus")
+        status_resource = provisioner_apigw_lambda.api_gateway.root.add_resource(
+            "pipelinestatus"
+        )
         status_resource.add_method("POST")
         blueprint_repository_bucket.grant_read(provisioner_apigw_lambda.lambda_function)
-        provisioner_apigw_lambda.lambda_function.add_to_role_policy(lambda_passrole_policy)
-        orchestrator_policy.attach_to_role(provisioner_apigw_lambda.lambda_function.role)
+        provisioner_apigw_lambda.lambda_function.add_to_role_policy(
+            lambda_passrole_policy
+        )
+        orchestrator_policy.attach_to_role(
+            provisioner_apigw_lambda.lambda_function.role
+        )
 
         # Environment variables setup
         provisioner_apigw_lambda.lambda_function.add_environment(
@@ -299,31 +365,47 @@ class MLOpsStack(core.Stack):
         provisioner_apigw_lambda.lambda_function.add_environment(
             key="ACCESS_BUCKET", value=str(access_logs_bucket.bucket_name)
         )
-        provisioner_apigw_lambda.lambda_function.add_environment(key="ASSETS_BUCKET", value=str(assets_s3_bucket_name))
+        provisioner_apigw_lambda.lambda_function.add_environment(
+            key="ASSETS_BUCKET", value=str(assets_s3_bucket_name)
+        )
         provisioner_apigw_lambda.lambda_function.add_environment(
             key="CFN_ROLE_ARN", value=str(cloudformation_role.role_arn)
         )
-        provisioner_apigw_lambda.lambda_function.add_environment(key="PIPELINE_STACK_NAME", value=pipeline_stack_name)
+        provisioner_apigw_lambda.lambda_function.add_environment(
+            key="PIPELINE_STACK_NAME", value=pipeline_stack_name
+        )
         provisioner_apigw_lambda.lambda_function.add_environment(
             key="NOTIFICATION_EMAIL", value=notification_email.value_as_string
         )
-        provisioner_apigw_lambda.lambda_function.add_environment(key="REGION", value=core.Aws.REGION)
-        provisioner_apigw_lambda.lambda_function.add_environment(key="IS_MULTI_ACCOUNT", value=str(multi_account))
+        provisioner_apigw_lambda.lambda_function.add_environment(
+            key="REGION", value=core.Aws.REGION
+        )
+        provisioner_apigw_lambda.lambda_function.add_environment(
+            key="IS_MULTI_ACCOUNT", value=str(multi_account)
+        )
         provisioner_apigw_lambda.lambda_function.add_environment(
             key="USE_MODEL_REGISTRY", value=use_model_registry.value_as_string
         )
         provisioner_apigw_lambda.lambda_function.add_environment(
-            key="ALLOW_DETAILED_ERROR_MESSAGE", value=allow_detailed_error_message.value_as_string
+            key="ALLOW_DETAILED_ERROR_MESSAGE",
+            value=allow_detailed_error_message.value_as_string,
         )
 
         provisioner_apigw_lambda.lambda_function.add_environment(
-            key="MLOPS_NOTIFICATIONS_SNS_TOPIC", value=mlops_notifications_topic.topic_arn
+            key="MLOPS_NOTIFICATIONS_SNS_TOPIC",
+            value=mlops_notifications_topic.topic_arn,
         )
-        provisioner_apigw_lambda.lambda_function.add_environment(key="ECR_REPO_NAME", value=ecr_repo_name)
+        provisioner_apigw_lambda.lambda_function.add_environment(
+            key="ECR_REPO_NAME", value=ecr_repo_name
+        )
 
-        provisioner_apigw_lambda.lambda_function.add_environment(key="ECR_REPO_ARN", value=ecr_repo_arn)
+        provisioner_apigw_lambda.lambda_function.add_environment(
+            key="ECR_REPO_ARN", value=ecr_repo_arn
+        )
 
-        provisioner_apigw_lambda.lambda_function.add_environment(key="LOG_LEVEL", value="DEBUG")
+        provisioner_apigw_lambda.lambda_function.add_environment(
+            key="LOG_LEVEL", value="DEBUG"
+        )
         cfn_policy_for_lambda = orchestrator_policy.node.default_child
         cfn_policy_for_lambda.cfn_options.metadata = {
             "cfn_nag": {
@@ -346,7 +428,9 @@ class MLOpsStack(core.Stack):
         repo_name_split = core.Fn.split("/", git_address.value_as_string)
         repo_name = core.Fn.select(5, repo_name_split)
         # getting codecommit repo cdk object using 'from_repository_name'
-        repo = codecommit.Repository.from_repository_name(self, "AWSMLOpsFrameworkRepository", repo_name)
+        repo = codecommit.Repository.from_repository_name(
+            self, "AWSMLOpsFrameworkRepository", repo_name
+        )
         codebuild_project = codebuild.PipelineProject(
             self,
             "Take config file",
@@ -398,10 +482,14 @@ class MLOpsStack(core.Stack):
             cross_account_keys=False,
         )
         codecommit_pipeline.add_to_role_policy(
-            create_invoke_lambda_policy([provisioner_apigw_lambda.lambda_function.function_arn])
+            create_invoke_lambda_policy(
+                [provisioner_apigw_lambda.lambda_function.function_arn]
+            )
         )
         codebuild_project.add_to_role_policy(
-            create_invoke_lambda_policy([provisioner_apigw_lambda.lambda_function.function_arn])
+            create_invoke_lambda_policy(
+                [provisioner_apigw_lambda.lambda_function.function_arn]
+            )
         )
         pipeline_child_nodes = codecommit_pipeline.node.find_all()
         pipeline_child_nodes[1].node.default_child.cfn_options.metadata = {
@@ -420,17 +508,25 @@ class MLOpsStack(core.Stack):
         }
 
         # custom resource for operational metrics###
-        metrics_mapping = core.CfnMapping(self, "AnonymousData", mapping={"SendAnonymousData": {"Data": "Yes"}})
+        metrics_mapping = core.CfnMapping(
+            self, "AnonymousData", mapping={"SendAnonymousData": {"Data": "Yes"}}
+        )
         metrics_condition = core.CfnCondition(
             self,
             "AnonymousDatatoAWS",
-            expression=core.Fn.condition_equals(metrics_mapping.find_in_map("SendAnonymousData", "Data"), "Yes"),
+            expression=core.Fn.condition_equals(
+                metrics_mapping.find_in_map("SendAnonymousData", "Data"), "Yes"
+            ),
         )
 
         # If user chooses Git as pipeline provision type, create codepipeline with Git repo as source
         core.Aspects.of(repo).add(ConditionalResources(git_address_provided))
-        core.Aspects.of(codecommit_pipeline).add(ConditionalResources(git_address_provided))
-        core.Aspects.of(codebuild_project).add(ConditionalResources(git_address_provided))
+        core.Aspects.of(codecommit_pipeline).add(
+            ConditionalResources(git_address_provided)
+        )
+        core.Aspects.of(codebuild_project).add(
+            ConditionalResources(git_address_provided)
+        )
 
         # Create Template Interface
         paramaters_list = [
@@ -444,11 +540,21 @@ class MLOpsStack(core.Stack):
         ]
 
         paramaters_labels = {
-            f"{notification_email.logical_id}": {"default": "Notification Email (Required)"},
-            f"{git_address.logical_id}": {"default": "CodeCommit Repo URL Address (Optional)"},
-            f"{existing_bucket.logical_id}": {"default": "Name of an Existing S3 Bucket (Optional)"},
-            f"{existing_ecr_repo.logical_id}": {"default": "Name of an Existing Amazon ECR repository (Optional)"},
-            f"{use_model_registry.logical_id}": {"default": "Do you want to use SageMaker Model Registry?"},
+            f"{notification_email.logical_id}": {
+                "default": "Notification Email (Required)"
+            },
+            f"{git_address.logical_id}": {
+                "default": "CodeCommit Repo URL Address (Optional)"
+            },
+            f"{existing_bucket.logical_id}": {
+                "default": "Name of an Existing S3 Bucket (Optional)"
+            },
+            f"{existing_ecr_repo.logical_id}": {
+                "default": "Name of an Existing Amazon ECR repository (Optional)"
+            },
+            f"{use_model_registry.logical_id}": {
+                "default": "Do you want to use SageMaker Model Registry?"
+            },
             f"{create_model_registry.logical_id}": {
                 "default": "Do you want the solution to create a SageMaker's model package group?"
             },
@@ -460,7 +566,11 @@ class MLOpsStack(core.Stack):
         # configure mutli-account parameters and permissions
         is_delegated_admin = None
         if multi_account:
-            paramaters_list, paramaters_labels, is_delegated_admin = configure_multi_account_parameters_permissions(
+            (
+                paramaters_list,
+                paramaters_labels,
+                is_delegated_admin,
+            ) = configure_multi_account_parameters_permissions(
                 self,
                 assets_bucket,
                 blueprint_repository_bucket,
@@ -489,7 +599,9 @@ class MLOpsStack(core.Stack):
             ).to_string(),
             "Region": core.Aws.REGION,
             "IsMultiAccount": str(multi_account),
-            "IsDelegatedAccount": is_delegated_admin if multi_account else core.Aws.NO_VALUE,
+            "IsDelegatedAccount": is_delegated_admin
+            if multi_account
+            else core.Aws.NO_VALUE,
             "UseModelRegistry": use_model_registry.value_as_string,
             "SolutionId": get_cdk_context_value(self, "SolutionId"),
             "Version": get_cdk_context_value(self, "Version"),