+++ CHANGELOG.md
@@ -5,12 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.2] - 2023-04-17
+
+### Updated
+
+- The S3 bucket policy on the logging S3 bucket to grant access to the logging service principal (logging.s3.amazonaws.com) for access log delivery.
+- Python libraries. 
+
 
 ## [2.1.1] - 2023-01-11
 
 ### Updated
 
-- Python runtime 3.10.
+- Python runtime 3.9.
 - Python libraries. 
 
 ## [2.1.0] - 2022-11-30